hw06-library-service-part2/src/main/java/ru/otus/spring/hw/repositories/CommentRepositoryJpa.java
package ru.otus.spring.hw.repositories;

import java.util.List;
import java.util.Optional;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

import org.springframework.stereotype.Repository;

import ru.otus.spring.hw.model.Comment;

@Repository
public class CommentRepositoryJpa implements CommentRepository {

    @PersistenceContext
    private EntityManager em;

    @Override
    public List<Comment> findAll() {
        return em.createQuery("select c from Comment c", Comment.class).getResultList();
    }

    @Override
    public Optional<Comment> findById(long id) {
        return Optional.ofNullable(em.find(Comment.class, id));
    }

    @Override
    public Comment save(Comment genre) {
        if (!genre.hasId()) {
            em.persist(genre);
            return genre;
        } else {
            return em.merge(genre);
        }
    }

    @Override
    public void remove(long id) {
        final var query = em.createQuery("delete from Comment c where c.id=:id");
        query.setParameter("id", id);
        final var result = query.executeUpdate();
        if (result == 0) {
            throw new RepositoryException(String.format("Comment with id %d was not deleted", id));
        }
    }

}
